Remove premature optimization
